+++ UniversityHW/Initial screens/LoginViewController.swift
@@ -24,45 +24,16 @@ class LoginViewController: UIViewController {
         let password = passwordTextField.text ?? ""
 
         //выполнение записи данных пользователя
-        let decodedUserInfo = fetchInfoFromFileStorage()
+        let decodedUserInfo = FileStorageManager.fetchInfoFromFileStorage()
         print(decodedUserInfo["username"] ?? "")
         let decodedUsername = decodedUserInfo["username"] ?? ""
         let decodedPassword = decodedUserInfo["password"] ?? ""
 
         if username == decodedUsername && password == decodedPassword {
             //запись данных в userDefaults
-            userDefaults()
+            MyUserDefaults.userDefaults()
             print("userDefaults saved")
             performSegue(withIdentifier: "fromLoginVCToWelcomeVC", sender: nil)
         }
-        else {
-        }
     }
-
-    //MARK: - File Storage (достаем данные пользователя)
-    func fetchInfoFromFileStorage() -> [String: String] {
-        var decodedUserInfo: [String : String] = [:]
-        let urls = FileManager.default.urls(for: .documentDirectory, in: .allDomainsMask)
-        if let fileURL = urls.first?.appendingPathComponent("users.txt") {
-            do {
-                let dataFromFile = try Data(contentsOf: fileURL)
-                let decoder = JSONDecoder()
-                decodedUserInfo = try decoder.decode([String: String].self, from: dataFromFile)
-                print(decodedUserInfo)
-            } catch {
-            }
-        }
-        return decodedUserInfo
-    }
-
-    //MARK: - User Defaults
-    func userDefaults() {
-        let userDefaults = UserDefaults.standard
-        //класть словари,массивы
-        userDefaults.set("true", forKey: "signed")
-        if let value = userDefaults.value(forKey: "signed") {
-            print(value)
-        }
-    }
-
 }
